[go] add configurations for production server
**Summary**

we decided to expose all the logic in production so this PR adds a configuration logic
to run http server on GCP environment while most of secrets are configured through datastore.

To run a server with production environment, the following command is required during the initial setup.

````
$ gcloud auth login
$ gcloud config set project {project}
$ export HPAPP_GCP_PROJECT_ID={project}
$ export HPAPP_GCP_DATASTORE_SETTINGS_KIND={kindname}
````

**Test**

- go run ./cmd/ --prod httpserver

then access `http://localhost:8080/graphql/v3/playground` to see if the GraphQL returns the production dataset

**Issue**

- N/A