[expo] add a setup to create a devclient build.
**Summary**

added a magical setup to kick the build process on top of EAS service.

we decided to use config/{config_name} as a directory to store all configurations for the particular EAS profile
and prevent developers from uploading some files like google-services.json. [*1]

we don't want to expose firebase api keys and some other secret. Those can be stored in config/{configname} during
the development workflow but should be uploaded to EAS secrets as a file type environment variable.

[*1] it's not much critical issue to commit to the repository (not necessary to reset the key if someone does it)
     but we don't want contributors to use that file for their development since it allows them to abuse
     the production GraphQL services. See also the discussion at Stackoverflow.

     https://stackoverflow.com/questions/37358340/should-i-add-the-google-services-json-from-firebase-to-my-repository

**Test**

$ yarn build-dev

confirmed the build is completed and it can be installed on my iOS device and running.

**Issue**

- N/A