[expo] add logging module to capture events locally
**Summary**

One of opportunities in v2.x seriese is about logging. We didn't have a good methodology to
get the application logs from the devices but we want to fix it for better triage process.

This PR adds a placeholder to capture logs and use console.log for the proof of concept.

We need to identify how to store logs on devices and upload to somewhere in the future.

**Test**

- expo

**Issue**

- N/A