[go][graphql] implement Node interface for HelloProject query
**Summary**

This is required to use pagination under HelloProject query. Otherwise, we see the following error on relay.

````
Failed to build:
 - Validation errors:
 - Invalid use of @refetchable on fragment 'FeedQuery_helloproject_query_feed', only supported are fragments on:
 - the Viewer type
 - the Query type
 - the Node interface, object types that implement the Node interface, interfaces whose implementing objects all implement Node, and unions whose members all implement Node
 - server objects and interfaces with the @fetchable directive:features/feed/Feed.tsx:12:45
````

We use `ID()` method signature to identify if the struct implements Node or not but Node interface is mapped to ent.Noder interface,
the struct has to implement `IsNode()` method as well.

**Test**

- go test

**Issue**

- N/A