[go][ent] add DESC index and use it in pagination
**Summary**

When we use Pagination feature in entql, we need to expicitly pass `ent.WithHPFeedItemOrder` option to sort the entities
instead of `Order()` function.

We also noticed there is no index of the combination of `post_at` and `id`, which should impact on the query performance.

Even with this chagne, tagged feed query is stil not optimized so we need to fix it in the future or shutdown the future.

**Test**

- go test

**Issue**

- #28