[expo][fix] myViewHistory field with schema merge
**Summary**

The schema files copied from the server side code uses `extend type`, which is interpreted
as a client schema extension so that relay don't generate a query string for extended fields
such as `myViewHistory`.

To solve this issue, we introduced `@graphql-tools` to merge the server schema files into single
schema file and use it for the client schema.

With the schema merge, the schema string like

```
type Foo {
    foo(): String!
}

extend type Foo {
    bar(): String!
}
```

is merged to

```
type Foo {
    foo(): String!
    bar(): String!
}
```

**Test**

- expo

**Issue**

- N/A