[expo] document authorization and app.config.js
**Summary**

We reviewed the authorization flow implementation and document it.

During the fresh development setup, we also noticed app.config.js logic is not clearly documented
and developers may waste time to troubleshoot configuration issues.

It was not clearly documented on EXPO/EAS side as well so we decided to document it in our project.
(Note if the behavior is changed, we should file a bug to EXPO/EAS)

**Test**

- expo

**Issue**

- N/A