[expo] introduce AppConfigModal
**Summary**

during the development, we sometimes want to use the different app config for the app without creating a new build.

so we stop to use expo-constants directly but use a wrapper with `useAppConfig()` hook.

At the initial boot, we copy necessary app config fields into local settings storage, and after that, we use the values in the local storage.

**Test**

- yarn test

**Issue**

- N/A