[expo] add storybook for react-native-web and integrate with jest
**Summary**

reuse storybook stories for react-native-web so that we can quickly confirm the look & feel on the web.

we also use the stories for jest automated testing so that we can easily detect regression of UI and also can extend the story tests to functionality tests.

this commit created `Platform.OS === "web"`` handling so we noted the web limitation on the issue #84

also go code is slightly changed to support CORS header so that storybook web can fetch the data from the go server.

**Test**

- yarn test

**Issue**

- #84 is related.