[go] refactor push notification device test tool.
**Summary**

We used to have a global settings of test notification tokens but now we open sourced the code so that
we should not use the global one - otherwise @yssk22 is always notified.

We introduced `TestTokens()` option instead of TesTokens settings in `Deliver` function so that developers can pass their own token.

**Test**

- go test

**Issue**

- N/A