[expo] update Android configurations
**Summary**

`dev.yssk22.hpapp` package is not available on Play Console as the developer account was suspended in early 2024.
I created a new developer account and use `dev.yssk22.hpapp.{env}.android` as a new package name. This change also
required new client ids for firebase authentication.

For Android, we don't need to use EAS server to build so we use `--local` to build the new android binary to
save the EAS cost.

**Test**

- `yarn build-android-dev`
- `yarn build-android-beta`

**Issue**

- #155