[expo] support debug token for AppCheck.
**Summary**

The development build is distributed not through AppStore but manual apk copy,
so it doesn't support AppCheck token with "PlayIntegrity" provider.

To make AppCheck available in the developmentn build, we use `debug` provider
to support AppCheck token for development build.

We also noticed that GraphQL endpoint is not customizable when starting metro server as we provided AppConfig modal to configure
but Dev Menu is not stable in Android build so that we moved it back to environment variable via `.env.local` file.

This change introduce the following environment variable:

- `APP_CHECK_DEBUG_TOKEN`
- `GRAPHQL_ENDPOINT`

And .env.local file is dynamically generate in Github Action workflow.

**Test**

- dev build

**Issue**

- #155