[expo] improve rendering logic
**Summary**

We noticed that #219 introduced the regression of HomeTab rendering and the root cause was
the 'rerendering' caused by useHomeTabContext() in HomeTab, which triggered 'unmount' of tabs
created by `Tabs.map()`

To avoid unmount while rerendering, we need to move the `Tabs.map()` to the fixed list of components instead of an array.

Also to ensure sections are properly cached by SectionList, we fixed keyExtractor logic, which can be moved to each `SectionListRender` class
so that React can evaluate the key by default instead of having an override `keyExtractor` function in the original SectionList.

**Test**

- expo

**Issue**

- N/A