[expo] update Android package name
**Summary**

dev.yssk22.hpapp.android was moved to the organization developer account but it was still being blocked by 12 testers policy.

Then I created a new app in the org dev account, it is not being blocked so it looks like the policy doesn't appliy for apps created by the org account.

To use apps in the org account, we need new package names.

**Test**

- N/A

**Issue**

- N/A