[expo] fix upsertFollow
**Summary**

we should not use getOrCreateLinkedRecord to create a new record, because it will create a new record even if the record already exists.
Instead, we should take a record proxy from the given HPFollow id, then set it to the linked record on HPArtist or HPMember
so that HPFollow and HPArtist/Member are fully linked and consistent.

**Test**

- expo

**Issue**

- #243